Restructure imports: package the tree, absolute imports from repo root
- Add empty __init__.py to smal_fitter/, smal_fitter/neuralSMIL/,
  smal_fitter/priors/, fitter_3d/, fitter_3d/pointcloud2smil/, custom_processing/
  (empty -> no import-time side effects; preserves the CUDA_VISIBLE_DEVICES-before-
  torch ordering the entrypoints rely on).
- Rename smal_fitter/smal_fitter.py -> smal_fitter/fitter.py to kill the
  dir==module ambiguity; SMALFitter now lives at smal_fitter.fitter.
- Rewrite every cross-module import to the absolute, repo-root-anchored form
  (smal_fitter.fitter, smal_fitter.utils, smal_fitter.neuralSMIL.<mod>,
  smal_fitter.priors.<mod>, fitter_3d.utils, ...). The two utils.py modules no
  longer collide.
- Remove all sys.path.append/insert hacks from application code.

Every module now imports identically regardless of CWD; entrypoints are launched
as `python -m <module>` from the repo root.
