+++ sleap_data_loader.py
@@ -17,8 +17,6 @@ import toml
 import pandas as pd
 import sys
 
-# Add the parent directories to the path to import config
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 import config
