+++ smal_fitter/sleap_data/optimized_sleap_preprocessor.py
@@ -24,16 +24,8 @@ import multiprocessing as mp
 from concurrent.futures import ProcessPoolExecutor, as_completed
 import json
 
-# Add paths for imports
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-# Import SLEAPDataLoader with fallback
-try:
-    from sleap_data_loader import SLEAPDataLoader
-except ImportError:
-    sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
-    from sleap_data_loader import SLEAPDataLoader
+from sleap_data_loader import SLEAPDataLoader
 
 import config
