+++ smal_fitter/sleap_data/sleap_multiview_dataset.py
@@ -15,9 +15,6 @@ from typing import Dict, List, Tuple, Optional, Any
 from pathlib import Path
 import sys
 
-# Add paths for imports
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 import config
